feat: cart user ;ogin

+++ src/pages/other/Cart.js
@@ -16,6 +16,7 @@ import {
 import LayoutOne from "../../layouts/LayoutOne";
 import Breadcrumb from "../../wrappers/breadcrumb/Breadcrumb";
 import Swal from "sweetalert2";
+import CartApi from "../../api/cart/CartApi";
 
 const Cart = ({
   location,
@@ -183,6 +184,22 @@ const Cart = ({
                                             confirmButtonText: "Tôi chắc chắn",
                                           }).then((result) => {
                                             if (result.isConfirmed) {
+                                              if (
+                                                JSON.parse(
+                                                  localStorage.getItem("user")
+                                                )
+                                              ) {
+                                                CartApi.ChangeCartItem({
+                                                  quantity: 1,
+                                                  userId: JSON.parse(
+                                                    localStorage.getItem("user")
+                                                  ).user.userId,
+                                                  productId: Number(
+                                                    cartItem.id
+                                                  ),
+                                                  IsAdd: 0,
+                                                });
+                                              }
                                               decreaseQuantity(
                                                 cartItem,
                                                 addToast
@@ -190,6 +207,20 @@ const Cart = ({
                                             }
                                           });
                                         } else {
+                                          if (
+                                            JSON.parse(
+                                              localStorage.getItem("user")
+                                            )
+                                          ) {
+                                            CartApi.ChangeCartItem({
+                                              quantity: 1,
+                                              userId: JSON.parse(
+                                                localStorage.getItem("user")
+                                              ).user.userId,
+                                              productId: Number(cartItem.id),
+                                              IsAdd: 0,
+                                            });
+                                          }
                                           decreaseQuantity(cartItem, addToast);
                                         }
                                       }}>
@@ -203,13 +234,27 @@ const Cart = ({
                                     />
                                     <button
                                       className="inc qtybutton"
-                                      onClick={() =>
+                                      onClick={() => {
+                                        if (
+                                          JSON.parse(
+                                            localStorage.getItem("user")
+                                          )
+                                        ) {
+                                          CartApi.ChangeCartItem({
+                                            quantity: 1,
+                                            userId: JSON.parse(
+                                              localStorage.getItem("user")
+                                            ).user.userId,
+                                            productId: Number(cartItem.id),
+                                            IsAdd: 1,
+                                          });
+                                        }
                                         addToCart(
                                           cartItem,
                                           addToast,
                                           quantityCount
-                                        )
-                                      }
+                                        );
+                                      }}
                                       disabled={
                                         cartItem !== undefined &&
                                         cartItem.quantity &&
@@ -248,9 +293,36 @@ const Cart = ({
 
                                 <td className="product-remove">
                                   <button
-                                    onClick={() =>
-                                      deleteFromCart(cartItem, addToast)
-                                    }>
+                                    onClick={() => {
+                                      Swal.fire({
+                                        title: "Bạn có muốn xóa không?",
+                                        text: "Bạn hãy suy nghĩ thật",
+                                        icon: "warning",
+                                        showCancelButton: true,
+                                        confirmButtonColor: "#3085d6",
+                                        cancelButtonColor: "#d33",
+                                        cancelButtonText: "Thôi",
+                                        confirmButtonText: "Tôi chắc chắn",
+                                      }).then((result) => {
+                                        if (result.isConfirmed) {
+                                          if (
+                                            JSON.parse(
+                                              localStorage.getItem("user")
+                                            )
+                                          ) {
+                                            CartApi.ChangeCartItem({
+                                              quantity: cartItem.quantity,
+                                              userId: JSON.parse(
+                                                localStorage.getItem("user")
+                                              ).user.userId,
+                                              productId: Number(cartItem.id),
+                                              IsAdd: 0,
+                                            });
+                                          }
+                                          deleteFromCart(cartItem, addToast);
+                                        }
+                                      });
+                                    }}>
                                     <i className="fa fa-times"></i>
                                   </button>
                                 </td>
@@ -285,6 +357,16 @@ const Cart = ({
                               confirmButtonText: "Đồng ý tôi đồng ý",
                             }).then((result) => {
                               if (result.isConfirmed) {
+                                if (result.isConfirmed) {
+                                  if (
+                                    JSON.parse(localStorage.getItem("user"))
+                                  ) {
+                                    CartApi.RemoveAllCart(
+                                      JSON.parse(localStorage.getItem("user"))
+                                        .user.userId
+                                    );
+                                  }
+                                }
                                 deleteAllFromCart(addToast);
                               }
                             });
